reverted bc breaks and added SecurityRolesMatrixType
