+++ UPGRADE-4.x.md
@@ -1,5 +1,5 @@
 UPGRADE 4.x
 ===========
 
-## Role Matrix
- The role matrix shows sonata admin permissions in a table. The roles hierarchy is still there
+## Role permission matrix
+The role permission matrix shows sonata admin permissions in a table. The roles hierarchy is still there
