+++ src/Resources/config/twig.xml
@@ -5,7 +5,7 @@
             <argument type="service" id="service_container"/>
         </service>
         <service id="sonata.user.twig.extensions" class="Sonata\UserBundle\Twig\SecurityExtension" public="false">
-            <argument type="service" id="sonata.user.editable_role_builder"/>
+            <argument type="service" id="sonata.user.roles_matrix_builder"/>
             <tag name="twig.extension"/>
         </service>
     </services>
