+++ src/Resources/views/Form/form_admin_fields.html.twig
@@ -10,6 +10,25 @@
  #}
 {% block sonata_security_roles_widget %}
     {% spaceless %}
+        <div class="editable">
+            <h4>{{ 'field.label_roles_editable'|trans({}, "SonataUserBundle") }}</h4>
+            {{ block('choice_widget') }}
+        </div>
+        {% if read_only_choices|length > 0 %}
+            <div class="readonly">
+                <h4>{{ 'field.label_roles_readonly'|trans({}, "SonataUserBundle") }}</h4>
+                <ul>
+                    {% for choice in read_only_choices %}
+                        <li>{{ choice }}</li>
+                    {% endfor %}
+                </ul>
+            </div>
+        {% endif %}
+    {% endspaceless %}
+{% endblock sonata_security_roles_widget %}
+
+{% block sonata_security_roles_matrix_widget %}
+    {% spaceless %}
         <table class="table table-condensed">
             <thead>
             <tr>
@@ -24,7 +43,7 @@
             </tbody>
         </table>
         <ul class="list-unstyled">
-           {{ renderCustomRolesList(form)|raw }}
+            {{ renderCustomRolesList(form)|raw }}
         </ul>
     {% endspaceless %}
-{% endblock sonata_security_roles_widget %}
+{% endblock sonata_security_roles_matrix_widget %}
